+++ mslicer/src/app/project.rs
@@ -144,7 +144,7 @@ impl<'a> BorrowedProject<'a> {
 
     pub fn serialize<Writer: Write>(&self, writer: &mut Writer) -> Result<()> {
         writer.write_all(&VERSION.to_le_bytes())?;
-        bincode::serialize_into(writer, self)?;
+        bincode::serde::encode_into_std_write(self, writer, bincode::config::standard())?;
         Ok(())
     }
 }
@@ -159,7 +159,10 @@ impl OwnedProject {
             anyhow::bail!("Invalid version: Expected {VERSION} found {version}");
         }
 
-        Ok(bincode::deserialize_from(reader)?)
+        Ok(bincode::serde::decode_from_std_read(
+            reader,
+            bincode::config::standard(),
+        )?)
     }
 
     pub fn apply(self, app: &mut App) {
@@ -167,7 +170,7 @@ impl OwnedProject {
         *meshes = self
             .meshes
             .into_iter()
-            .map(|mesh| mesh.into_rendered_mesh())
+            .map(|mesh| mesh.into_rendered_mesh(app))
             .collect();
 
         app.slice_config = self.slice_config;
@@ -175,17 +178,19 @@ impl OwnedProject {
 }
 
 impl OwnedProjectMesh {
-    pub fn into_rendered_mesh(self) -> RenderedMesh {
+    pub fn into_rendered_mesh(self, app: &App) -> RenderedMesh {
         let mut mesh = Mesh::new_uncentred(self.vertices, self.faces, self.normals);
         mesh.set_position_unchecked(self.info.position);
         mesh.set_scale_unchecked(self.info.scale);
         mesh.set_rotation_unchecked(self.info.rotation);
         mesh.update_transformation_matrix();
 
-        RenderedMesh::from_mesh(mesh)
+        let mut rendered = RenderedMesh::from_mesh(mesh)
             .with_name(self.info.name)
             .with_color(self.info.color)
-            .with_hidden(self.info.hidden)
+            .with_hidden(self.info.hidden);
+        rendered.update_oob(&app.slice_config);
+        rendered
     }
 }