+++ mslicer/src/render/pipelines/target_point.rs
@@ -71,17 +71,19 @@ impl TargetPointPipeline {
             layout: Some(&pipeline_layout),
             vertex: VertexState {
                 module: &shader,
-                entry_point: "vert",
+                entry_point: None,
                 buffers: &[VERTEX_BUFFER_LAYOUT],
+                compilation_options: Default::default(),
             },
             fragment: Some(FragmentState {
                 module: &shader,
-                entry_point: "frag",
+                entry_point: None,
                 targets: &[Some(ColorTargetState {
                     format: texture,
                     blend: Some(BlendState::ALPHA_BLENDING),
                     write_mask: ColorWrites::all(),
                 })],
+                compilation_options: Default::default(),
             }),
             primitive: PrimitiveState::default(),
             depth_stencil: Some(DepthStencilState {
@@ -96,6 +98,7 @@ impl TargetPointPipeline {
                 ..Default::default()
             },
             multiview: None,
+            cache: None,
         });
 
         let (vertices, indices) = generate_sphere(20);
@@ -143,11 +146,7 @@ impl TargetPointPipeline {
         queue.write_buffer(&self.uniform_buffer, 0, &buffer.into_inner());
     }
 
-    pub fn paint<'a>(
-        &'a self,
-        render_pass: &mut RenderPass<'a>,
-        resources: &WorkspaceRenderCallback,
-    ) {
+    pub fn paint(&self, render_pass: &mut RenderPass, resources: &WorkspaceRenderCallback) {
         if !resources.is_moving {
             return;
         };