+++ mslicer/src/windows/models.rs
@@ -1,14 +1,20 @@
+use common::misc::subscript_number;
 use const_format::concatcp;
 use egui::{Context, Grid, Id, Ui};
 use egui_phosphor::regular::{
-    ARROW_LINE_DOWN, COPY, DICE_THREE, EYE, EYE_SLASH, LINK_BREAK, LINK_SIMPLE, TRASH,
+    ARROW_LINE_DOWN, CIRCLES_THREE, COPY, DICE_THREE, EYE, EYE_SLASH, LINK_BREAK, LINK_SIMPLE,
+    TRASH, TRIANGLE, WARNING,
 };
 use slicer::Pos;
 
 use crate::{
     app::App,
+    render::rendered_mesh::MeshWarnings,
     ui::components::{vec3_dragger, vec3_dragger_proportional},
 };
+
+const WARN_NON_MANIFOLD: &str = "This mesh is non-manifold, it may produce unexpected results when sliced.\nConsider running it through a mesh repair tool.";
+const WARN_OUT_OF_BOUNDS: &str = "This mesh extends beyond the printer volume and will be cut off.";
 
 enum Action {
     None,
@@ -42,6 +48,19 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
                 .on_hover_text(if mesh.hidden { "Show" } else { "Hide" })
                 .clicked();
 
+            if !mesh.warnings.is_empty() {
+                let count = mesh.warnings.bits().count_ones();
+                let mut warn = ui.label(format!("{WARNING}{}", subscript_number(count)));
+                for warning in mesh.warnings.iter() {
+                    let desc = match warning {
+                        MeshWarnings::NonManifold => WARN_NON_MANIFOLD,
+                        MeshWarnings::OutOfBounds => WARN_OUT_OF_BOUNDS,
+                        _ => unreachable!(),
+                    };
+                    warn = warn.on_hover_text(desc);
+                }
+            }
+
             if open {
                 ui.text_edit_singleline(&mut mesh.name);
             } else {
@@ -54,6 +73,18 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
                 .num_columns(2)
                 .with_row_color(|row, style| (row % 2 == 0).then_some(style.visuals.faint_bg_color))
                 .show(ui, |ui| {
+                    ui.label("Info");
+                    ui.horizontal(|ui| {
+                        ui.label(TRIANGLE);
+                        ui.monospace(mesh.mesh.face_count().to_string());
+
+                        ui.separator();
+
+                        ui.label(CIRCLES_THREE);
+                        ui.monospace(mesh.mesh.vertex_count().to_string());
+                    });
+                    ui.end_row();
+
                     ui.label("Actions");
                     ui.vertical(|ui| {
                         ui.horizontal_wrapped(|ui| {
@@ -65,7 +96,10 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
                                 .then(|| action = Action::Duplicate(i));
                             ui.button(concatcp!(ARROW_LINE_DOWN, " Align to Bed"))
                                 .clicked()
-                                .then(|| mesh.align_to_bed());
+                                .then(|| {
+                                    mesh.align_to_bed();
+                                    mesh.update_oob(&app.slice_config);
+                                });
                         });
                     });
                     ui.end_row();
@@ -78,7 +112,7 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
                         let mut position = mesh.mesh.position();
                         vec3_dragger(ui, position.as_mut(), |x| x);
                         (mesh.mesh.position() != position)
-                            .then(|| mesh.mesh.set_position(position));
+                            .then(|| mesh.set_position(app, position));
                         ui.add_space(width);
                     });
                     ui.end_row();
@@ -89,14 +123,14 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
                         let mut scale = mesh.mesh.scale();
                         if mesh.locked_scale {
                             vec3_dragger_proportional(ui, scale.as_mut(), |x| {
-                                x.speed(0.01).clamp_range(0.001..=f32::MAX)
+                                x.speed(0.01).range(0.001..=f32::MAX)
                             });
                         } else {
                             vec3_dragger(ui, scale.as_mut(), |x| {
-                                x.speed(0.01).clamp_range(0.001..=f32::MAX)
+                                x.speed(0.01).range(0.001..=f32::MAX)
                             });
                         }
-                        (mesh.mesh.scale() != scale).then(|| mesh.mesh.set_scale(scale));
+                        (mesh.mesh.scale() != scale).then(|| mesh.set_scale(app, scale));
 
                         mesh.locked_scale ^= ui
                             .button(if mesh.locked_scale {
@@ -113,7 +147,7 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
                     let original_rotation = rotation;
                     vec3_dragger(ui, rotation.as_mut(), |x| x.suffix("°"));
                     (original_rotation != rotation)
-                        .then(|| mesh.mesh.set_rotation(deg_to_rad(rotation)));
+                        .then(|| mesh.set_rotation(app, deg_to_rad(rotation)));
                     ui.end_row();
 
                     ui.label("Color");
@@ -126,14 +160,6 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
 
                     ui.label("Name");
                     ui.text_edit_singleline(&mut mesh.name);
-                    ui.end_row();
-
-                    ui.label("Vertices");
-                    ui.monospace(mesh.mesh.vertex_count().to_string());
-                    ui.end_row();
-
-                    ui.label("Triangles");
-                    ui.monospace(mesh.mesh.face_count().to_string());
                     ui.end_row();
                 });
         }