+++ mslicer/src/windows/slice_config.rs
@@ -39,7 +39,13 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
             ui.end_row();
 
             ui.label("Platform Size (mm)");
-            vec3_dragger(ui, app.slice_config.platform_size.as_mut(), |x| x);
+            let platform = &mut app.slice_config.platform_size;
+            let prev = *platform;
+            vec3_dragger(ui, platform.as_mut(), |x| x);
+            if *platform != prev {
+                (app.meshes.write().iter_mut())
+                    .for_each(|model| model.update_oob(&app.slice_config));
+            }
             ui.end_row();
 
             ui.label("Slice Height (mm)");
@@ -84,23 +90,23 @@ fn exposure_config_grid(ui: &mut Ui, config: &mut ExposureConfig) {
         .striped(true)
         .show(ui, |ui| {
             ui.label("Exposure Time (s)");
-            ui.add(DragValue::new(&mut config.exposure_time).clamp_range(0.0..=f32::MAX));
+            ui.add(DragValue::new(&mut config.exposure_time).range(0.0..=f32::MAX));
             ui.end_row();
 
             ui.label("Lift Distance (mm)");
-            ui.add(DragValue::new(&mut config.lift_distance).clamp_range(0.0..=f32::MAX));
+            ui.add(DragValue::new(&mut config.lift_distance).range(0.0..=f32::MAX));
             ui.end_row();
 
             ui.label("Lift Speed (cm/min)");
-            ui.add(DragValue::new(&mut config.lift_speed).clamp_range(0.0..=f32::MAX));
+            ui.add(DragValue::new(&mut config.lift_speed).range(0.0..=f32::MAX));
             ui.end_row();
 
             ui.label("Retract Distance (mm)");
-            ui.add(DragValue::new(&mut config.retract_distance).clamp_range(0.0..=f32::MAX));
+            ui.add(DragValue::new(&mut config.retract_distance).range(0.0..=f32::MAX));
             ui.end_row();
 
             ui.label("Retract Speed (cm/min)");
-            ui.add(DragValue::new(&mut config.retract_speed).clamp_range(0.0..=f32::MAX));
+            ui.add(DragValue::new(&mut config.retract_speed).range(0.0..=f32::MAX));
             ui.end_row();
         });
 }