+++ mslicer/src/windows/workspace.rs
@@ -1,6 +1,5 @@
 use const_format::concatcp;
-use eframe::Theme;
-use egui::{ComboBox, Context, DragValue, Grid, Ui};
+use egui::{ComboBox, Context, DragValue, Grid, Theme, Ui};
 use egui_phosphor::regular::{ARROW_COUNTER_CLOCKWISE, FOLDER, LAYOUT};
 use tracing::error;
 
@@ -53,7 +52,7 @@ pub fn ui(app: &mut App, ui: &mut Ui, _ctx: &Context) {
         });
 
     dragger(ui, "Grid Size", &mut app.config.grid_size, |x| {
-        x.speed(0.1).clamp_range(1.0..=f32::MAX)
+        x.speed(0.1).range(1.0..=f32::MAX)
     });
 
     ui.add_space(16.0);