+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.5.0 &mdash; Coming Soon™
+
+- Detect non-manifold meshes
+- Color backfaces grey in rendered view
+
 ## v0.4.0 &mdash; November 25th, 2025
 
 - Added support for the encrypted Chitu format (.ctb)